requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
anthropic>=0.49.0
SQLAlchemy==2.0.25
aiosqlite==0.19.0
aiofiles==23.2.1
Pillow==10.1.0
python-dateutil==2.8.2
openai>=1.14.0
yookassa>=3.0.0
